+++ build.gradle.kts
@@ -54,7 +54,6 @@ dependencies {
 	implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:2.2.0")
 	implementation("me.paulschwarz:spring-dotenv:4.0.0")
 	implementation("com.querydsl:querydsl-jpa:$queryDslVersion:jakarta")
-	implementation("org.springframework.boot:spring-boot-starter-security")
 	kapt("com.querydsl:querydsl-apt:$queryDslVersion:jakarta")
 }
