+++ src/main/kotlin/com/blife/blife/domain/book/repository/LibBookRepository.kt
@@ -1,0 +1,8 @@
+package com.blife.blife.domain.book.repository
+
+import com.blife.blife.domain.book.model.LibBook
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface LibBookRepository : JpaRepository<LibBook, Long> {
+    fun findByBookIdAndLibId(bookId : Long, libraryId : Long) : LibBook?
+}
