+++ src/main/kotlin/com/blife/blife/domain/review/dto/AverageScoreDto.kt
@@ -1,0 +1,5 @@
+package com.blife.blife.domain.review.dto
+
+data class AverageScoreDto(
+    var point : Double
+)
